+++ src/containers/delete-playlist-confirmation-modal/components/DeletePlaylistConfirmationModal.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 
+import ActionSheetModal from 'components/action-sheet-modal/ActionDrawer'
+
 import { ID } from 'common/models/Identifiers'
-import ActionSheetModal from 'components/action-sheet-modal/ActionSheetModal'
 
 type DeletePlaylistConfirmationModalProps = {
   isOpen: boolean
