+++ src/containers/unfollow-confirmation-modal/components/UnfollowConfirmationModal.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 
+import ActionSheetModal from 'components/action-sheet-modal/ActionDrawer'
+
 import { ID } from 'common/models/Identifiers'
-import ActionSheetModal from 'components/action-sheet-modal/ActionSheetModal'
 
 type UnfollowConfirmationModalProps = {
   isOpen: boolean
